refactor(orderable-bytes): rename `numeric` module to `primitive`
The module now hosts a `bool` impl alongside the integer and float
impls; `primitive` describes the contents more accurately than
`numeric`.
